Add default RouteServiceProvider
This can be overwritten on a per-application basis in `config/app.php`.
Its purpose is to load the routes in `routes/api.php` for browsing.

I'm adding it here, because we would be using the same exact file
for all of our microservices. This reduces code duplication across
our repos. I hope it's not too unintuitive, that a new provider can
be created for a given application and used in place of this one.