Add --connection option to db:reset [WEB-133]
User data needs to be stored in tables that do not have the same
versioned prefixes as normal resource tables. We need to have the
ability to version our database without wiping out user data. To
do that, we'll use a constant `auth_` prefix. We can't leave it
unprefixed, or `db:reset` will wipe out all tables. With this
change, we can now specify which connection to target for each
call of `db:reset` as defined in `config/database.php`.