Add `ForceAcceptJson` middleware [WEB-1516]
This makes it so that Laravel's internal `wantsJson` returns true,
which is really only relevant to exception handling, for the most part.
But in either case, regardless of what `Accept` headers the client asks
for from our API, we will *always* return JSON, so adding this to the
stack seems like a good idea.